moved fucntions to assign and remove users from surveys to user.ts
