+++ src/app/api/surveys/remove/route.tsx
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { removeSurveys } from '@/lib/googleForms';
+import { removeSurveys } from '@/lib/firebase/database/users'; 
 
 export async function POST(req: NextRequest) {
     try {
